added support for busybox lsusb
dymodes decides a mode-switch scheme to follow based on name of the
chosen device, and it gets this listing of usb devices from the 'lsusb'
tool.  however, busybox's lsusb doesn't print name of a device.  i've
improved the software to identify manufacturer of the device based on
it's vendor ID on busybox-lsusb distros.  in that way, dymodes will also
work on distros using busybox's lsusb (which are often times tiny
distros).